ISSUE-8: code refactor
