+++ client/Makefile
@@ -1,0 +1,6 @@
+all:
+	pyinstaller -F main.py -n ClientApp
+
+default_target:
+	all
+.PHONY: default_target
