+++ client/commands/__init__.py
@@ -1,0 +1,5 @@
+from .command import Command
+from .bad_command import BadCommand
+from .show_help import ShowHelp
+from .send_string import SendString
+from .exit import ExitCommand
