+++ client/commands/bad_command.py
@@ -1,0 +1,6 @@
+class BadCommand(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
