+++ environment.txt
@@ -1,0 +1,85 @@
+name: base
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - altgraph=0.17=pyhd3eb1b0_0
+  - arrow=1.1.1=pyhd8ed1ab_0
+  - beautifulsoup4=4.9.3=pyha847dfd_0
+  - binaryornot=0.4.4=py_1
+  - brotlipy=0.7.0=py39h89e85a6_1001
+  - bzip2=1.0.8=h1de35cc_0
+  - ca-certificates=2021.7.5=hecd8cb5_1
+  - certifi=2021.5.30=py39hecd8cb5_0
+  - cffi=1.14.6=py39hb71fe58_0
+  - chardet=4.0.0=py39h6e9494a_1
+  - charset-normalizer=2.0.0=pyhd8ed1ab_0
+  - click=8.0.1=py39h6e9494a_0
+  - conda=4.10.3=py39hecd8cb5_0
+  - conda-build=3.21.4=py39hecd8cb5_0
+  - conda-package-handling=1.7.3=py39h9ed2024_1
+  - cookiecutter=1.7.3=pyh6c4a22f_0
+  - cryptography=3.4.7=py39ha2c9959_0
+  - filelock=3.0.12=pyhd3eb1b0_1
+  - glob2=0.7=pyhd3eb1b0_0
+  - gmp=6.2.1=h23ab428_2
+  - icu=58.2=h0a44026_3
+  - idna=3.1=pyhd3deb0d_0
+  - jinja2=2.11.3=pyh44b312d_0
+  - jinja2-time=0.2.0=py_2
+  - libarchive=3.4.2=haa3ed63_0
+  - libcxx=10.0.0=1
+  - libffi=3.3=hb1e8313_2
+  - libiconv=1.16=h1de35cc_0
+  - liblief=0.10.1=h23ab428_1
+  - libxml2=2.9.12=hcdb78fc_0
+  - lz4-c=1.9.3=h23ab428_1
+  - macholib=1.14=pyhd3eb1b0_1
+  - markupsafe=1.1.1=py39hcbf5805_3
+  - ncurses=6.2=h0a44026_1
+  - openssl=1.1.1k=h9ed2024_0
+  - pip=21.2.4=py37hecd8cb5_0
+  - pkginfo=1.7.1=py39hecd8cb5_0
+  - poyo=0.5.0=py_0
+  - psutil=5.8.0=py39h9ed2024_1
+  - py-lief=0.10.1=py39h23ab428_1
+  - pycosat=0.6.3=py39h9ed2024_0
+  - pycparser=2.20=pyh9f0ad1d_2
+  - pycryptodome=3.10.1=py39h659dea6_0
+  - pyinstaller=3.6=py39h7bc2e8c_6
+  - pyopenssl=20.0.1=pyhd8ed1ab_0
+  - pysocks=1.7.1=py39h6e9494a_3
+  - python=3.9.6=h88f2d9e_1
+  - python-dateutil=2.8.2=pyhd8ed1ab_0
+  - python-libarchive-c=2.9=pyhd3eb1b0_1
+  - python-slugify=5.0.2=pyhd8ed1ab_0
+  - python_abi=3.9=2_cp39
+  - pytz=2021.1=pyhd3eb1b0_0
+  - pyyaml=5.4.1=py39h9ed2024_1
+  - readline=8.1=h9ed2024_0
+  - requests=2.26.0=pyhd8ed1ab_0
+  - ripgrep=12.1.1=0
+  - ruamel_yaml=0.15.100=py39h9ed2024_0
+  - setuptools=52.0.0=py39hecd8cb5_0
+  - six=1.16.0=pyh6c4a22f_0
+  - soupsieve=2.2.1=pyhd3eb1b0_0
+  - sqlite=3.36.0=hce871da_0
+  - text-unidecode=1.3=py_0
+  - tk=8.6.10=hb0a8c7a_0
+  - tqdm=4.62.1=pyhd3eb1b0_1
+  - typing_extensions=3.10.0.0=pyha770c72_0
+  - tzdata=2021a=h5d7bf9c_0
+  - unidecode=1.2.0=pyhd8ed1ab_0
+  - urllib3=1.26.6=pyhd8ed1ab_0
+  - wheel=0.37.0=pyhd3eb1b0_0
+  - xz=5.2.5=h1de35cc_0
+  - yaml=0.2.5=haf1e3a3_0
+  - zlib=1.2.11=h1de35cc_3
+  - zstd=1.4.9=h322a384_0
+  - pip:
+    - build==0.6.0.post1
+    - packaging==21.0
+    - pep517==0.11.0
+    - pyparsing==2.4.7
+    - tomli==1.2.1
+prefix: /Users/a.shestakov/PycharmProjects/NetworkBasicApplication/env
