+++ libs/common_lib/common_lib/exceptions.py
@@ -1,0 +1,2 @@
+class ExitException(Exception):
+    ...
