+++ libs/common_lib/common_lib/utilities.py
@@ -1,0 +1,5 @@
+def delete_multiple_element(list_object, indices):
+    indices = sorted(indices, reverse=True)
+    for idx in indices:
+        if idx < len(list_object):
+            list_object.pop(idx)
