+++ libs/common_lib/setup.py
@@ -1,0 +1,13 @@
+from setuptools import setup
+
+setup(
+    name='common_lib',
+    version='0.0.1',
+    author='ShestA',
+    author_email='6akov98@gmail.com',
+    packages=['common_lib'],
+    url='https://github.com/ShestA/NetworkBasicApplication/',
+    description='Common utilities for NetworkBasicApplication',
+    install_requires=[
+    ],
+)
