+++ libs/network_lib/Makefile
@@ -1,0 +1,7 @@
+all:
+	python3 setup.py build
+	python3 setup.py develop
+.PHONY: all
+
+default_target: all
+.PHONY: default_target
