+++ libs/network_lib/setup.py
@@ -1,0 +1,13 @@
+from setuptools import setup
+
+setup(
+    name='network_lib',
+    version='0.0.1',
+    author='ShestA',
+    author_email='6akov98@gmail.com',
+    packages=['network_lib'],
+    url='https://github.com/ShestA/NetworkBasicApplication/',
+    description='Utilities for NetworkBasicApplication',
+    install_requires=[
+    ],
+)
