+++ server/Makefile
@@ -1,0 +1,6 @@
+all:
+	pyinstaller -F main.py -n ServerApp
+
+default_target:
+	all
+.PHONY: default_target
